refactor: improve highlighted function (#4967)
